Move the coordinate transformation tests to test_transform_layer.py
`test_dataset.py` had grown a block of tests that are about the transformation classes
themselves rather than about datasets, so they move to `test_transform_layer.py`
together with their two helpers.

`test_layer_coordinate_transformations` used `prepare_dataset_path` and
`assure_exported_properties`, which are module-level helpers of `test_dataset.py`. No
test module imports from another one, so it now takes the `tmp_upath` fixture like every
other test in its new home and asserts the round-trip through the file inline.

The two remaining mentions of `coordinate_transformations` in `test_dataset.py` stay
where they are: one asserts on the parsed `LayerProperties`, the other covers what
copying a layer carries over.
